+++ example/android/app/src/main/java/com/igaworks/adbrixrm_flutter_example/MainActivity.java
@@ -1,6 +1,23 @@
 package com.igaworks.adbrixrm_flutter_example;
 
+import android.content.Intent;
+import android.os.Bundle;
+
 import io.flutter.embedding.android.FlutterActivity;
 
 public class MainActivity extends FlutterActivity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState){
+        super.onCreate(savedInstanceState);
+        onNewIntent(this.getIntent());
+    }
+
+    @Override
+    protected void onNewIntent(Intent intent) {
+        super.onNewIntent(intent);
+        setIntent(intent);
+
+    }
+
+
 }